Padrino::Cache::Helpers::* moved from padrino-cache
