+++ padrino-cache-helpers/lib/padrino-cache-helpers/cache_store.rb
