+++ padrino-cache-helpers/lib/padrino-cache-helpers/fragment.rb
