+++ padrino-cache-helpers/lib/padrino-cache-helpers/page.rb
