+++ padrino-cache/lib/padrino-cache.rb
@@ -1,6 +1,5 @@
 require 'fileutils'
 require 'padrino-core'
-require 'padrino-helpers'
 FileSet.glob_require('padrino-cache/{helpers}/*.rb', __FILE__)
 
 module Padrino
@@ -58,14 +57,6 @@ module Padrino
 
     class << self
       ##
-      # Register these helpers:
-      #
-      #   Padrino::Cache::Helpers::CacheStore
-      #   Padrino::Cache::Helpers::Fragment
-      #   Padrino::Cache::Helpers::Page
-      #
-      # for Padrino::Application.
-      #
       # By default we use FileStore as showed below:
       #
       #   set :cache, Padrino::Cache::Store::File.new(File.join(app.root, 'tmp', 'cache'))
@@ -88,18 +79,10 @@ module Padrino
       #
       # @api public
       def registered(app)
-        app.helpers Padrino::Cache::Helpers::CacheStore
-        app.helpers Padrino::Cache::Helpers::Fragment
-        app.helpers Padrino::Cache::Helpers::Page
         app.set :cache, Padrino::Cache::Store::File.new(Padrino.root('tmp', app.app_name.to_s ,'cache'))
         app.disable :caching
       end
       alias :included :registered
-
-      # @private
-      def padrino_route_added(route, verb, path, args, options, block) # @private
-        Padrino::Cache::Helpers::Page.padrino_route_added(route, verb, path, args, options, block)
-      end
     end
   end # Cache
 end # Padrino
